Split activity.rs into focused modules
Refactors the 1149-line activity.rs file into a modular structure:

- activity/models.rs: Type definitions (InputType, InputContext, AgentInput,
  AgentOutput, AgentMetric, TokenUsage, ActivityEntry, ProductivitySummary)
- activity/schema.rs: Database schema SQL and init_schema() function
- activity/db.rs: ActivityDb struct and all database operations
- activity/mod.rs: Module coordination and public exports

This improves:
- Maintainability: Each module has single responsibility
- Navigation: Developers can find code faster
- Code review: PRs touching activity logic are easier to review
- Testing: Modules can be tested in isolation

All existing tests pass. Public API unchanged.

Closes #783
